requirements.txt
openai-whisper
torch
ffmpeg-python
python-telegram-bot
